# omod/src/main/resources/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "REST API for SMS Module.",
    "version" : "v1",
    "title" : "SMS Module REST API"
  },
  "basePath" : "/openmrs/ws",
...
  "paths" : {
    "/scheduleAdHocSMS" : {
      "post" : {
        "tags" : [ "REST API to schedule AdHoc SMSes based on data from different input sources" ],
        "summary" : "None",
        "description" : "Schedules AdHoc SMSes based on data from file",
        "operationId" : "scheduleAdHocSMSesFromFile",
        "parameters" : [ {
          "in" : "body",
          "name" : "file",
          "description" : "file",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/MultipartFile"
          }
        }, {
          "name" : "sheetName",
          "in" : "query",
          "description" : "sheetName",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "AdHoc SMSes have been successfully scheduled"
          },
          "500" : {
            "description" : "Failed to schedule AdHoc SMSes"
          }
        }
      }
    },
    "/scheduleAdHocSMS/{dataSetUuid}" : {
      "post" : {
        "tags" : [ "REST API to schedule AdHoc SMSes based on data from different input sources" ],
        "summary" : "None",
        "description" : "Schedules AdHoc SMSes based on data from SQL data set",
        "operationId" : "scheduleAdHocSMSesFromDataSet",
        "parameters" : [ {
          "name" : "dataSetUuid",
          "in" : "path",
          "description" : "dataSetUuid",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "AdHoc SMSes have been successfully scheduled"
          },
          "500" : {
            "description" : "Failed to schedule AdHoc SMSes"
          }
        }
      }
    },
...
  },
  "definitions" : {
...
    "InputStream" : {
      "type" : "object"
    },
    "MultipartFile" : {
      "type" : "object",
      "properties" : {
        "originalFilename" : {
          "type" : "string"
        },
        "contentType" : {
          "type" : "string"
        },
        "contentType" : {
          "type" : "string"
        },
        "name" : {
          "type" : "string"
        },
        "empty" : {
          "type" : "boolean",
          "default" : false
        },
        "bytes" : {
          "type" : "array",
          "items" : {
            "type" : "string",
            "format" : "byte"
          }
        },
        "size" : {
          "type" : "integer",
          "format" : "int64"
        },
        "inputStream" : {
          "$ref" : "#/definitions/InputStream"
        }
      }
    },
...
  }
}